% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addTable}
\alias{addTable}
\title{Add table to existing SQLite database.}
\usage{
addTable(db_path = "", table_name = "", tab_file = "")
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{table_name}{Name for new table}

\item{tab_file}{Tab-delimited file to add to database}
}
\value{
NULL (modified SQLite database)
}
\description{
\code{addTable} adds a table to an existing
SQLite database from a tab-delimited file,
e.g., Microbes Online genome annotation file
}
\examples{
addTable( db_path = "/path/to/db", table_new = "my_new_table", tab_file = "path/to/tab/delim/file" )
}

