% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/getBiclusterExpression.R
\name{getBiclusterExpression}
\alias{getBiclusterExpression}
\title{Get bicluster expression values from SQLite database.}
\usage{
getBiclusterExpression(db_path = "", bicluster = 1, iteration = "max",
  ratios = "")
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}

\item{ratios}{Path to tab-delimited file or variable containing gene expression data}
}
\value{
Matrix containing gene expression values for each gene and condition in the bicluster
}
\description{
\code{getBiclusterExpression} retrieves expression values (genes, conditions) from a
user specified bicluster and/or motif number at any iteration
of cMonkey algorithm
}

