% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getConditions}
\alias{getConditions}
\title{Get bicluster conditions from SQLite database.}
\usage{
getConditions(db_path = "", bicluster = 1, iteration = "max")
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}
}
\value{
Vector of conditions in the bicluster
}
\description{
\code{getConditions} retrieves conditions from a
user specified bicluster at any iteration
of cMonkey algorithm
}
\examples{
getConditions( db_path = "/path/to/database", bicluster = 10)
getConditions( db_path = "/path/to/database", bicluster = 100, iteration = 1002)
}

