% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGenes}
\alias{getGenes}
\title{Get bicluster genes from SQLite database.}
\usage{
getGenes(db_path = "", bicluster = 1, iteration = "max")
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}
}
\value{
Vector of genes in the bicluster
}
\description{
\code{getGenes} retrieves genes from a
user specified bicluster at any iteration
of cMonkey algorithm
}
\examples{
getGenes( db_path = "/path/to/database", bicluster = 10)
getGenes( db_path = "/path/to/database", bicluster = 100, iteration = 1002)
}

