% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMemeMotifs}
\alias{getMemeMotifs}
\title{Get bicluster MEME motif info from SQLite database.}
\usage{
getMemeMotifs(db_path = "", bicluster = 1, iteration = "max",
  motif_num = seq(1, 10))
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}

\item{motif_num}{Number of the motif to return. Defaults to all motifs
(assuming # motifs per bicluster < 10)}
}
\value{
data.frame containing information about the locations and properties of discovered motifs by MEME
}
\description{
\code{getgetMemeMotifs} retrieves MEME output  from a
user specified bicluster and/or motif number at any iteration
of cMonkey algorithm
}
\examples{
getMemeMotifs( db_path = "/path/to/database", bicluster = 10)
getMemeMotifs( db_path = "/path/to/database", bicluster = 100, iteration = 1002)
getMemeMotifs( db_path = "/path/to/database", bicluster = 100, iteration = 1002, motif_num = 2)
}

