% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMotifPWM}
\alias{getMotifPWM}
\title{Get bicluster motif PWMs from SQLite database.}
\usage{
getMotifPWM(db_path = "", bicluster = 1, iteration = "max",
  motif_num = seq(1, 10))
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}

\item{motif_num}{Number of the motif to return. Defaults to all motifs
(assuming # motifs per bicluster < 10)}
}
\value{
List containing $pwm and $evalue for each motif requested
}
\description{
\code{getMotifPWM} retrieves motif PWM(s) from a
user specified bicluster and/or motif number at any iteration
of cMonkey algorithm
}
\examples{
getMotifPWM( db_path = "/path/to/database", bicluster = 10)
getMotifPWM( db_path = "/path/to/database", bicluster = 100, iteration = 1002)
getMotifPWM( db_path = "/path/to/database", bicluster = 100, iteration = 1002, motif_num = 2)
}

