% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/load_cMonkey.R
\name{load_cMonkey}
\alias{load_cMonkey}
\title{Load cMonkey database.}
\usage{
load_cMonkey(path = "", verbose = T)
}
\arguments{
\item{path}{Location of the SQLite database}

\item{verbose}{Should location of database be written to console}
}
\value{
Connection to SQLite database
}
\description{
\code{load_cMonkey} Loads the SQLite databases produced from a
cMonkey2 (aka python cMonkey) run
}

