% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/plotMotif.R
\name{plotMotif}
\alias{plotMotif}
\title{Plot cMonkey PSSM using seqLogo.}
\usage{
plotMotif(pwm = NULL, db_path = "", bicluster = 1, iteration = "max",
  motif_num = 1, ic.scale = TRUE, file = NULL)
}
\arguments{
\item{pwm}{A PWM of the form:}

\item{pwm}{A PWM of the form:}
}
\value{
Plot of the motif
}
\description{
\code{plotMotif} plots a PWM
}

