% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{Get bicluster genes from SQLite database.}
\usage{
getGenes(db_path = "", bicluster = 1, iteration = "max")
}
\arguments{
\item{db_path}{Location of the SQLite database}

\item{bicluster}{Bicluster number, integer}

\item{iteration}{Iteration of bicluster to retrieve, defaults to last (final) iteration}
}
\value{
Vector of genes in the bicluster
}
\description{
\code{getGenes} retrieves genes from a
user specified bicluster at any iteration
of cMonkey algorithm
}

